AutomationToolkit/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "openai>=1.97.0",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pyautogui>=0.9.54",
    "requests>=2.32.4",
]
